Fixed library for nodejs Webrtc Client
I am exploring a system where I would like to use Webrtc on both the
browser clients and the server.

I expect to have a small number of servers that will interact with the
system just like any other client, only they will have slightly better
uptime and persistence capabilities.

Unfortunately, the y-webrtc library loads the browser only build of
simple-peer, and this causes a failure when run inside of a node
application.

This change does not work (as-is) when used to build for the browser
(using esbuild) because simple-peer requires nodejs events module, which
esbuild doesn't want to put into the browser.

Because it breaks the browser case, please do not pull this change
upstream.